[Bug Fix] Fixed cursor position when merging lines by backspace
Bug was related to the way in which the merge point was calculated.
It relied on actual characters in the buffer,
rather than on their visual representation.
When line contained tabs, it resulted in the visual cursor position shifting
tab_len-1 columns to the left for each tab in the line.

Solution is to store visual end of the previous line before merge,
move to the beginning after the merge and iteratively simulate pressing
right key until the cursor reaches stored position.

It may cause performance issues but it's the best solution for now.